refactor!: standardize directory structure and namespace (#3, #4)
